+++ pennylane/ops/qubit.py
@@ -1914,7 +1914,7 @@ class QubitUnitary(Operation):
         return U
 
     def adjoint(self):
-        return QubitUnitary(qml.math.T(qml.math.conj(self.data[0])), wires=self.wires)
+        return QubitUnitary(qml.math.T(qml.math.conj(self.matrix)), wires=self.wires)
 
     def _controlled(self, wire):
         ControlledQubitUnitary(*self.parameters, control_wires=wire, wires=self.wires)
@@ -1987,6 +1987,7 @@ class ControlledQubitUnitary(QubitUnitary):
             raise ValueError(f"Input unitary must be of shape {(target_dim, target_dim)}")
 
         # Saving for the circuit drawer
+        self._target_wires = wires
         self.control_wires = control_wires
         self.U = U
 
@@ -1997,6 +1998,7 @@ class ControlledQubitUnitary(QubitUnitary):
             control_values = "1" * len(control_wires)
 
         control_int = self._parse_control_values(control_wires, control_values)
+        self.control_values = control_values
 
         # A multi-controlled operation is a block-diagonal matrix partitioned into
         # blocks where the operation being applied sits in the block positioned at
@@ -2006,15 +2008,19 @@ class ControlledQubitUnitary(QubitUnitary):
         # etc. The positioning of the block is controlled by padding the block diagonal
         # to the left and right with the correct amount of identity blocks.
 
-        padding_left = control_int * len(U)
-        padding_right = 2 ** len(wires) - len(U) - padding_left
-
-        CU = block_diag(np.eye(padding_left), U, np.eye(padding_right))
-
-        params = list(params)
-        params[0] = CU
+        self._padding_left = control_int * len(U)
+        self._padding_right = 2 ** len(wires) - len(U) - self._padding_left
+        self._CU = None
 
         super().__init__(*params, wires=wires, do_queue=do_queue)
+
+    def _matrix(self, *params):
+        if self._CU is None:
+            self._CU = block_diag(np.eye(self._padding_left), self.U, np.eye(self._padding_right))
+
+        params = list(params)
+        params[0] = self._CU
+        return super()._matrix(*params)
 
     @staticmethod
     def _parse_control_values(control_wires, control_values):
@@ -2034,7 +2040,8 @@ class ControlledQubitUnitary(QubitUnitary):
         return control_int
 
     def _controlled(self, wire):
-        ControlledQubitUnitary(*self.parameters, control_wires=wire, wires=self.wires)
+        ctrl_wires = sorted(self.control_wires + wire)
+        ControlledQubitUnitary(*self.parameters, control_wires=ctrl_wires, wires=self._target_wires)
 
 
 class MultiControlledX(ControlledQubitUnitary):
@@ -2052,6 +2059,31 @@ class MultiControlledX(ControlledQubitUnitary):
         wires (Union[Wires or int]): a single target wire the operation acts on
         control_values (str): a string of bits representing the state of the control
             qubits to control on (default is the all 1s state)
+        work_wires (Union[Wires, Sequence[int], or int]): optional work wires used to decompose
+            the operation into a series of Toffoli gates
+
+    .. note::
+
+        If ``MultiControlledX`` is not supported on the targeted device, PennyLane will decompose
+        the operation into :class:`~.Toffoli` and/or :class:`~.CNOT` gates. When controlling on
+        three or more wires, the Toffoli-based decompositions described in Lemmas 7.2 and 7.3 of
+        `Barenco et al. <https://arxiv.org/abs/quant-ph/9503016>`__ will be used. These methods
+        require at least one work wire.
+
+        The number of work wires provided determines the decomposition method used and the resulting
+        number of Toffoli gates required. When ``MultiControlledX`` is controlling on :math:`n`
+        wires:
+
+        #. If at least :math:`n - 2` work wires are provided, the decomposition in Lemma 7.2 will be
+           applied using the first :math:`n - 2` work wires.
+        #. If fewer than :math:`n - 2` work wires are provided, a combination of Lemmas 7.3 and 7.2
+           will be applied using only the first work wire.
+
+        These methods present a tradeoff between qubit number and depth. The method in point 1
+        requires fewer Toffoli gates but a greater number of qubits.
+
+        Note that the state of the work wires before and after the decomposition takes place is
+        unchanged.
 
     **Example**
 
@@ -2070,9 +2102,24 @@ class MultiControlledX(ControlledQubitUnitary):
     par_domain = "A"
     grad_method = None
 
-    def __init__(self, control_wires=None, wires=None, control_values=None, do_queue=True):
-        if len(Wires(wires)) != 1:
+    # pylint: disable=too-many-arguments
+    def __init__(
+        self, control_wires=None, wires=None, control_values=None, work_wires=None, do_queue=True
+    ):
+        wires = Wires(wires)
+        control_wires = Wires(control_wires)
+        work_wires = Wires([]) if work_wires is None else Wires(work_wires)
+
+        if len(wires) != 1:
             raise ValueError("MultiControlledX accepts a single target wire.")
+
+        if Wires.shared_wires([wires, work_wires]) or Wires.shared_wires(
+            [control_wires, work_wires]
+        ):
+            raise ValueError("The work wires must be different from the control and target wires")
+
+        self._target_wire = wires[0]
+        self._work_wires = work_wires
 
         super().__init__(
             np.array([[0, 1], [1, 0]]),
@@ -2081,6 +2128,113 @@ class MultiControlledX(ControlledQubitUnitary):
             control_values=control_values,
             do_queue=do_queue,
         )
+
+    # pylint: disable=unused-argument
+    def decomposition(self, *args, **kwargs):
+
+        if len(self.control_wires) > 2 and len(self._work_wires) == 0:
+            raise ValueError(f"At least one work wire is required to decompose operation: {self}")
+
+        flips1 = [
+            qml.PauliX(self.control_wires[i])
+            for i, val in enumerate(self.control_values)
+            if val == "0"
+        ]
+
+        if len(self.control_wires) == 1:
+            decomp = [qml.CNOT(wires=[self.control_wires[0], self._target_wire])]
+        elif len(self.control_wires) == 2:
+            decomp = [qml.Toffoli(wires=[*self.control_wires, self._target_wire])]
+        else:
+            num_work_wires_needed = len(self.control_wires) - 2
+
+            if len(self._work_wires) >= num_work_wires_needed:
+                decomp = self._decomposition_with_many_workers(
+                    self.control_wires, self._target_wire, self._work_wires
+                )
+            else:
+                work_wire = self._work_wires[0]
+                decomp = self._decomposition_with_one_worker(
+                    self.control_wires, self._target_wire, work_wire
+                )
+
+        flips2 = [
+            qml.PauliX(self.control_wires[i])
+            for i, val in enumerate(self.control_values)
+            if val == "0"
+        ]
+
+        return flips1 + decomp + flips2
+
+    @staticmethod
+    def _decomposition_with_many_workers(control_wires, target_wire, work_wires):
+        """Decomposes the multi-controlled PauliX gate using the approach in Lemma 7.2 of
+        https://arxiv.org/pdf/quant-ph/9503016.pdf, which requires a suitably large register of
+        work wires"""
+        num_work_wires_needed = len(control_wires) - 2
+        work_wires = work_wires[:num_work_wires_needed]
+
+        work_wires_reversed = list(reversed(work_wires))
+        control_wires_reversed = list(reversed(control_wires))
+
+        gates = []
+
+        for i in range(len(work_wires)):
+            ctrl1 = control_wires_reversed[i]
+            ctrl2 = work_wires_reversed[i]
+            t = target_wire if i == 0 else work_wires_reversed[i - 1]
+            gates.append(qml.Toffoli(wires=[ctrl1, ctrl2, t]))
+
+        gates.append(qml.Toffoli(wires=[*control_wires[:2], work_wires[0]]))
+
+        for i in reversed(range(len(work_wires))):
+            ctrl1 = control_wires_reversed[i]
+            ctrl2 = work_wires_reversed[i]
+            t = target_wire if i == 0 else work_wires_reversed[i - 1]
+            gates.append(qml.Toffoli(wires=[ctrl1, ctrl2, t]))
+
+        for i in range(len(work_wires) - 1):
+            ctrl1 = control_wires_reversed[i + 1]
+            ctrl2 = work_wires_reversed[i + 1]
+            t = work_wires_reversed[i]
+            gates.append(qml.Toffoli(wires=[ctrl1, ctrl2, t]))
+
+        gates.append(qml.Toffoli(wires=[*control_wires[:2], work_wires[0]]))
+
+        for i in reversed(range(len(work_wires) - 1)):
+            ctrl1 = control_wires_reversed[i + 1]
+            ctrl2 = work_wires_reversed[i + 1]
+            t = work_wires_reversed[i]
+            gates.append(qml.Toffoli(wires=[ctrl1, ctrl2, t]))
+
+        return gates
+
+    @staticmethod
+    def _decomposition_with_one_worker(control_wires, target_wire, work_wire):
+        """Decomposes the multi-controlled PauliX gate using the approach in Lemma 7.3 of
+        https://arxiv.org/pdf/quant-ph/9503016.pdf, which requires a single work wire"""
+        tot_wires = len(control_wires) + 2
+        partition = int(np.ceil(tot_wires / 2))
+
+        first_part = control_wires[:partition]
+        second_part = control_wires[partition:]
+
+        gates = [
+            MultiControlledX(
+                control_wires=first_part, wires=work_wire, work_wires=second_part + target_wire
+            ),
+            MultiControlledX(
+                control_wires=second_part + work_wire, wires=target_wire, work_wires=first_part
+            ),
+            MultiControlledX(
+                control_wires=first_part, wires=work_wire, work_wires=second_part + target_wire
+            ),
+            MultiControlledX(
+                control_wires=second_part + work_wire, wires=target_wire, work_wires=first_part
+            ),
+        ]
+
+        return gates
 
 
 class DiagonalQubitUnitary(DiagonalOperation):